+++ sources/plugins/MailSender/MailSender.cpp
@@ -1,36 +1,33 @@
 #include "stdafx.h"
 #include "MailSender.h"
-#include <Poco/Net/MailMessage.h>
+
+#include <regex>
+#include <curlpp/Infos.hpp>
 #include <shared/plugin/yPluginApi/historization/Message.h>
 #include <shared/exception/EmptyResult.hpp>
 #include <plugin_cpp_api/ImplementationHelper.h>
 #include "MSConfiguration.h"
-#include "SmtpServiceProviderFactory.h"
 #include <shared/Log.h>
-#include "CertificatePassphraseProvider.h"
+#include <shared/http/curlppHelpers.h>
+#include <curlpp/Options.hpp>
 
-// Use this macro to define all necessary to make your DLL a Yadoms valid plugin.
-// Note that you have to provide some extra files, like package.json, and icon.png
-// This macro also defines the static PluginInformations value that can be used by plugin to get information values
+#include "MailBodyBuilder.h"
 
 IMPLEMENT_PLUGIN(CMailSender)
+
+/// @brief Recipient field used to retrieve e-mail from a recipient
+static const std::string MailRecipientField("email");
 
 CMailSender::CMailSender() :
    m_deviceName("MailSender"),
    m_configuration(boost::make_shared<CMSConfiguration>()),
-   m_mailId("email"),
-   m_messageKeyword(boost::make_shared<yApi::historization::CMessage>("message", m_mailId, yApi::EKeywordAccessMode::kGetSet)),
-   m_certificatePassphraseProvider(new CCertificatePassphraseProvider(m_configuration))
-   {
-}
-
-CMailSender::~CMailSender()
+   m_messageKeyword(boost::make_shared<yApi::historization::CMessage>("message", MailRecipientField, yApi::EKeywordAccessMode::kGetSet))
 {
 }
 
 void CMailSender::doWork(boost::shared_ptr<yApi::IYPluginApi> api)
 {
-   YADOMS_LOG(information) << "MailSender is starting..." ;
+   YADOMS_LOG(information) << "MailSender is starting...";
 
 
    m_configuration->initializeWith(api->getConfiguration());
@@ -38,37 +35,40 @@ void CMailSender::doWork(boost::shared_ptr<yApi::IYPluginApi> api)
    declareDevice(api);
 
    // the main loop
-   YADOMS_LOG(information) << "MailSender plugin is running..." ;
+   YADOMS_LOG(information) << "MailSender plugin is running...";
+   api->setPluginState(yApi::historization::EPluginState::kRunning);
 
-   while (1)
+   while (true)
    {
       // Wait for an event
       switch (api->getEventHandler().waitForEvents())
       {
       case yApi::IYPluginApi::kEventStopRequested:
          {
-            YADOMS_LOG(information) << "Stop requested" ;
+            YADOMS_LOG(information) << "Stop requested";
             api->setPluginState(yApi::historization::EPluginState::kStopped);
             return;
          }
       case yApi::IYPluginApi::kEventUpdateConfiguration:
          {
+            api->setPluginState(yApi::historization::EPluginState::kCustom, "updateConfiguration");
             onUpdateConfiguration(api, api->getEventHandler().getEventData<boost::shared_ptr<shared::CDataContainer>>());
+            api->setPluginState(yApi::historization::EPluginState::kRunning);
             break;
          }
 
       case yApi::IYPluginApi::kEventDeviceCommand:
          {
             // Command received
-            auto command = api->getEventHandler().getEventData<boost::shared_ptr<const yApi::IDeviceCommand>>();
-            YADOMS_LOG(information) << "Command received :" << yApi::IDeviceCommand::toString(command) ;
+            const auto command = api->getEventHandler().getEventData<boost::shared_ptr<const yApi::IDeviceCommand>>();
+            YADOMS_LOG(information) << "Command received :" << yApi::IDeviceCommand::toString(command);
 
             try
             {
                if (boost::iequals(command->getKeyword(), m_messageKeyword->getKeyword()))
                   onSendMailRequest(api, command->getBody());
                else
-               YADOMS_LOG(information) << "Received command for unknown keyword from Yadoms : " << yApi::IDeviceCommand::toString(command) ;
+                  YADOMS_LOG(information) << "Received command for unknown keyword from Yadoms : " << yApi::IDeviceCommand::toString(command);
             }
             catch (std::exception& e)
             {
@@ -98,7 +98,7 @@ void CMailSender::onUpdateConfiguration(boost::shared_ptr<yApi::IYPluginApi> api
                                         const boost::shared_ptr<shared::CDataContainer>& newConfigurationData) const
 {
    // Configuration was updated
-   YADOMS_LOG(information) << "Configuration was updated..." ;
+   YADOMS_LOG(information) << "Configuration was updated...";
    BOOST_ASSERT(!newConfigurationData->empty()); // newConfigurationData shouldn't be empty, or kEventUpdateConfiguration shouldn't be generated
 
    // Update configuration
@@ -106,36 +106,61 @@ void CMailSender::onUpdateConfiguration(boost::shared_ptr<yApi::IYPluginApi> api
 }
 
 void CMailSender::onSendMailRequest(boost::shared_ptr<yApi::IYPluginApi> api,
-                                    const std::string& sendMailRequest)
+                                    const std::string& sendMailRequest) const
 {
    try
    {
       m_messageKeyword->setCommand(sendMailRequest);
 
-      std::string from = m_configuration->getSenderEmail();
-      std::string to = getRecipientMail(api, m_messageKeyword->to());
+      const auto from = m_configuration->getSenderEmail();
+      const auto to = api->getRecipientValue(m_messageKeyword->to(), MailRecipientField);
 
-      std::string subject = "Yadoms Notification !";
-      Poco::Net::MailMessage message;
-      message.setSender(from);
-      message.addRecipient(Poco::Net::MailRecipient(Poco::Net::MailRecipient::PRIMARY_RECIPIENT, to));
-      message.setSubject(subject);
-      message.setContentType("text/plain; charset=UTF-8");
-      message.setContent(m_messageKeyword->body(), Poco::Net::MailMessage::ENCODING_8BIT);
+      curlpp::Easy mailSendRequest;
+      mailSendRequest.setOpt(new curlpp::options::Url(m_configuration->getHost()));
+      mailSendRequest.setOpt(new curlpp::options::Port(m_configuration->getPort()));
+      mailSendRequest.setOpt(new curlpp::options::MailFrom(from));
+      mailSendRequest.setOpt(new curlpp::options::MailRcpt({to}));
 
-      boost::shared_ptr<ISmtpServiceProvider> mailServiceProvider = CSmtpServiceProviderFactory::CreateSmtpServer(m_configuration, m_certificatePassphraseProvider);
-      if (mailServiceProvider->sendMail(message))
+      const auto bodyString = CMailBodyBuilder(from,
+                                               {to})
+                              .setSubject(formatSubject(m_messageKeyword->body()))
+                              .setAsciiBody(m_messageKeyword->body())
+                              .build();
+
+      size_t bytesRead = 0;
+      mailSendRequest.setOpt(new curlpp::options::ReadFunction([&bodyString, &bytesRead](char* ptr, size_t size, size_t nbItems)
       {
-         YADOMS_LOG(information) << "Email successfully sent to " << to ;
-      }
-      else
+         if (size == 0 || nbItems == 0 || size * nbItems < 1 || bytesRead > bodyString.size())
+            return static_cast<size_t>(0);
+
+         const size_t len = std::min(bodyString.size() - bytesRead, size * nbItems);
+         memcpy(ptr, &bodyString.c_str()[bytesRead], len);
+         bytesRead += len;
+
+         return len;
+      }));
+      mailSendRequest.setOpt(new curlpp::options::Upload(true));
+
+      setSecurity(mailSendRequest);
+      setAuthentication(mailSendRequest);
+
+      try
       {
-         YADOMS_LOG(error) << "Fail to send email to " << to ;
+         mailSendRequest.perform();
       }
+      catch (const curlpp::LibcurlRuntimeError& error)
+      {
+         const auto message = (boost::format("Fail to send email : %1%, code %2%") % error.what() % error.whatCode()).str();
+         throw std::runtime_error(message);
+      }
+
+      YADOMS_LOG(information) << "Email successfully sent to " << to;
+
+      //TODO gérer le proxy
    }
    catch (shared::exception::CInvalidParameter& e)
    {
-      YADOMS_LOG(error) << "Invalid Mail sending request \"" << sendMailRequest << "\" : " << e.what() ;
+      YADOMS_LOG(error) << "Invalid Mail sending request \"" << sendMailRequest << "\" : " << e.what();
    }
    catch (std::exception& e)
    {
@@ -143,21 +168,66 @@ void CMailSender::onSendMailRequest(boost::shared_ptr<yApi::IYPluginApi> api,
    }
    catch (...)
    {
-      YADOMS_LOG(error) << "Error sending Mail" ;
+      YADOMS_LOG(error) << "Error sending Mail";
    }
 }
 
 
-std::string CMailSender::getRecipientMail(boost::shared_ptr<yApi::IYPluginApi> api,
-                                          int recipientId) const
+void CMailSender::setSecurity(curlpp::Easy& mailSendRequest) const
 {
-   try
+   switch (m_configuration->getSecurityMode())
    {
-      return api->getRecipientValue(recipientId, m_mailId);
-   }
-   catch (shared::exception::CEmptyResult& e)
-   {
-      throw shared::exception::CInvalidParameter(std::string("Recipient not found, or e-mail not found for recipient : ") + e.what());
+   case ESecurityMode::kNoneValue:
+      mailSendRequest.setOpt(new curlpp::options::UseSsl(CURLUSESSL_NONE));
+      return;
+   case ESecurityMode::kSSLValue:
+      mailSendRequest.setOpt(new curlpp::options::UseSsl(CURLUSESSL_TRY));
+      mailSendRequest.setOpt(new curlpp::options::SslVerifyHost(2));
+#ifdef WIN32
+      mailSendRequest.setOpt(curlpp::options::SslOptions(CURLSSLOPT_NATIVE_CA));
+#endif
+      return;
+   case ESecurityMode::kTLSValue:
+      mailSendRequest.setOpt(new curlpp::options::UseSsl(CURLUSESSL_ALL));
+      mailSendRequest.setOpt(new curlpp::options::SslVerifyHost(2));
+#ifdef WIN32
+      mailSendRequest.setOpt(curlpp::options::SslOptions(CURLSSLOPT_NATIVE_CA));
+#endif
+      return;
+   case ESecurityMode::kAutomaticValue:
+      mailSendRequest.setOpt(new curlpp::options::SslVerifyHost(2));
+#ifdef WIN32
+      mailSendRequest.setOpt(curlpp::options::SslOptions(CURLSSLOPT_NATIVE_CA));
+#endif
+      switch (m_configuration->getPort())
+      {
+      case 465: //default ssl port
+         mailSendRequest.setOpt(new curlpp::options::UseSsl(CURLUSESSL_TRY));
+         return;
+      case 587: //default tsl port
+         mailSendRequest.setOpt(new curlpp::options::UseSsl(CURLUSESSL_ALL));
+         return;
+      default: //25, 26 or other
+         mailSendRequest.setOpt(new curlpp::options::UseSsl(CURLUSESSL_TRY));
+         return;
+      }
+   default:
+      throw std::invalid_argument("Unknown security mode " + std::to_string(m_configuration->getSecurityMode()));
    }
 }
 
+void CMailSender::setAuthentication(curlpp::Easy& mailSendRequest) const
+{
+   if (!m_configuration->getAuthenticationRequired())
+      return;
+
+   mailSendRequest.setOpt(curlpp::options::UserPwd(m_configuration->getLogin() + ":" + m_configuration->getPassword()));
+}
+
+std::string CMailSender::formatSubject(const std::string& text) const
+{
+   const auto firstTextLine = text.substr(0, text.find_first_of("\r\n"));
+   return std::regex_replace(m_configuration->getMailSubject(),
+                             std::regex("{m}", std::regex_constants::ECMAScript | std::regex_constants::extended),
+                             firstTextLine);
+}